+++ dashboard/modules/snapshot/tests/conftest.py
@@ -1,1 +1,1 @@
-from ray._private.test_utils import set_override_dashboard_url  # noqa: F401
+from ray._private.conftest_utils import set_override_dashboard_url  # noqa: F401
